Add pixel dim feature
Main problem was that the style property of element refers only to the
inline styles. But there was a case when the style was computed from the
external css file and returned empty string for the background color
property. That's why I needed to also check computed styles of window
object.